refactor: split validation to numbers and common
